catch_ai/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.1
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
psycopg2-binary
djangorestframework-simplejwt
python-decouple
dj-database-url
nanoid
gunicorn
google-auth
requests
firebase-admin